refactor: offer sent slider
