+++ src/modules/Contract/containers/SliderOfferSent/useSliderOfferSent.tsx
@@ -1,0 +1,28 @@
+import { useDispatch, useSelector } from 'react-redux';
+import { cancelContractAdaptor, Contract } from 'src/core/adaptors';
+import { CurrentIdentity } from 'src/core/api';
+import { RootState } from 'src/store';
+import { updateStatus } from 'src/store/reducers/contracts.reducer';
+
+export const useSliderOfferSent = (contract: Contract) => {
+  const dispatch = useDispatch();
+  const identity = useSelector<RootState, CurrentIdentity | undefined>(state =>
+    state.identity.entities.find(identity => identity.current),
+  );
+  const identityType = identity?.type;
+
+  const onWithdrawOffer = async () => {
+    await cancelContractAdaptor(contract.id);
+    dispatch(
+      updateStatus({
+        id: contract.id,
+        status: 'PROVIDER_CANCELED',
+        isCurrentProvider: identityType === contract.providerId,
+        type: contract.type,
+        paymentId: contract.paymentId,
+      }),
+    );
+  };
+
+  return { operations: { onWithdrawOffer } };
+};
